backend/requirements.txt
google-genai>=0.3.0
python-dotenv>=1.0.0
asyncio
websockets>=12.0
aiohttp>=3.9.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0 